Cleanup spec directory
There was some wonkiness that's bothered me for a while: an empty file,
the safe statements tests in the wrong filename, and so on. I didn't
change any tests here, just cleaned up where they live.
