--- spec/safe_migrations_spec.rb
@@ -1,7 +1,0 @@
-require "spec_helper"
-
-RSpec.describe PgHaMigrations do
-  it "has a version number" do
-    expect(PgHaMigrations::VERSION).not_to be nil
-  end
-end
